Refactor blocks + hide non-rendered jump-to sections
