+++ app/assets/javascripts/3dbio_viewer/src/webapp/components/protvista/JumpToButton.tsx
@@ -2,17 +2,20 @@ import React from "react";
 import _ from "lodash";
 import i18n from "../../utils/i18n";
 import { Dropdown, DropdownProps } from "../dropdown/Dropdown";
-import { blockDefs } from "./protvista-blocks";
+import { BlockDef } from "./Protvista.types";
 
-export interface JumpToButtonProps {}
+export interface JumpToButtonProps {
+    blocks: BlockDef[];
+}
 
-export const JumpToButton: React.FC<JumpToButtonProps> = () => {
+export const JumpToButton: React.FC<JumpToButtonProps> = React.memo(props => {
+    const { blocks } = props;
     const items: DropdownProps["items"] = React.useMemo(() => {
-        return blockDefs.map(block => ({ id: block.id, text: block.title }));
-    }, []);
+        return blocks.map(block => ({ id: block.id, text: block.title }));
+    }, [blocks]);
 
     return <Dropdown text={i18n.t("Jump to")} items={items} onClick={goToElement} />;
-};
+});
 
 function goToElement(DOMElementId: string) {
     // Use document.getElementById for simplicity. The orthodox approach would be to use refs,
