+++ app/assets/javascripts/3dbio_viewer/src/webapp/components/protvista/Protvista.types.ts
@@ -1,4 +1,5 @@
 import React from "react";
+import _ from "lodash"
 import { Pdb } from "../../../domain/entities/Pdb";
 import { TrackDefinition } from "../../../domain/entities/TrackDefinition";
 import { PdbView } from "../../view-models/PdbView";
@@ -31,4 +32,17 @@ export interface BlockDef extends ViewerBlockModel {
 export interface ProtvistaBlock extends ViewerBlockModel {
     tracks: TrackDef[];
     component?: React.FC<BlockComponentProps>;
+}
+
+export function blockHasRelevantData(block: BlockDef, pdb: Pdb): boolean {
+    const tracks = _(pdb.tracks)
+        .keyBy(track => track.id)
+        .at(...block.tracks.map(trackDef => trackDef.id))
+        .compact()
+        .value();
+    const trackIds = tracks.map(track => track.id);
+    const hasCustomComponent = Boolean(block.component);
+    const hasRelevantTracks = !_(tracks).isEmpty() && !_.isEqual(trackIds, ["structure-coverage"]);
+
+    return hasCustomComponent || hasRelevantTracks;
 }
